+++ react_on_rails_pro/spec/react_on_rails_pro/react_on_rails_pro_spec.rb
@@ -4,6 +4,6 @@ require_relative "spec_helper"
 
 describe ReactOnRailsPro do
   it "has a version number" do
-    expect(ReactOnRailsPro::VERSION).not_to be nil
+    expect(ReactOnRailsPro::VERSION).not_to be_nil
   end
 end